+++ example_test.go
@@ -16,341 +16,139 @@ limitations under the License.
 package oras_test
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
 	"fmt"
-	"net/http"
-	"net/http/httptest"
-	"net/url"
-	"os"
-	"strconv"
-	"strings"
-	"testing"
 
-	"github.com/opencontainers/go-digest"
-	specs "github.com/opencontainers/image-spec/specs-go"
-	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
+	v1 "github.com/opencontainers/image-spec/specs-go/v1"
 	"oras.land/oras-go/v2"
-	"oras.land/oras-go/v2/content/memory"
+	"oras.land/oras-go/v2/content/file"
 	"oras.land/oras-go/v2/content/oci"
-	"oras.land/oras-go/v2/internal/spec"
 	"oras.land/oras-go/v2/registry/remote"
+	"oras.land/oras-go/v2/registry/remote/auth"
+	"oras.land/oras-go/v2/registry/remote/retry"
 )
 
-var exampleMemoryStore oras.Target
-var remoteHost string
-var (
-	exampleManifest, _ = json.Marshal(spec.Artifact{
-		MediaType:    spec.MediaTypeArtifactManifest,
-		ArtifactType: "example/content"})
-	exampleManifestDescriptor = ocispec.Descriptor{
-		MediaType: spec.MediaTypeArtifactManifest,
-		Digest:    digest.Digest(digest.FromBytes(exampleManifest)),
-		Size:      int64(len(exampleManifest))}
-	exampleSignatureManifest, _ = json.Marshal(spec.Artifact{
-		MediaType:    spec.MediaTypeArtifactManifest,
-		ArtifactType: "example/signature",
-		Subject:      &exampleManifestDescriptor})
-	exampleSignatureManifestDescriptor = ocispec.Descriptor{
-		MediaType: spec.MediaTypeArtifactManifest,
-		Digest:    digest.FromBytes(exampleSignatureManifest),
-		Size:      int64(len(exampleSignatureManifest))}
-)
-
-func pushBlob(ctx context.Context, mediaType string, blob []byte, target oras.Target) (desc ocispec.Descriptor, err error) {
-	desc = ocispec.Descriptor{ // Generate descriptor based on the media type and blob content
-		MediaType: mediaType,
-		Digest:    digest.FromBytes(blob), // Calculate digest
-		Size:      int64(len(blob)),       // Include blob size
+func Example_pullFilesFromRemoteRepository() {
+	// 0. Create a file store
+	fs, err := file.New("/tmp/")
+	if err != nil {
+		panic(err)
 	}
-	return desc, target.Push(ctx, desc, bytes.NewReader(blob)) // Push the blob to the registry target
-}
+	defer fs.Close()
 
-func generateManifestContent(config ocispec.Descriptor, layers ...ocispec.Descriptor) ([]byte, error) {
-	content := ocispec.Manifest{
-		Config:    config, // Set config blob
-		Layers:    layers, // Set layer blobs
-		Versioned: specs.Versioned{SchemaVersion: 2},
-	}
-	return json.Marshal(content) // Get json content
-}
-
-func TestMain(m *testing.M) {
-	const exampleTag = "latest"
-	const exampleUploadUUid = "0bc84d80-837c-41d9-824e-1907463c53b3"
-
-	// Setup example local target
-	exampleMemoryStore = memory.New()
-	layerBlob := []byte("Hello layer")
+	// 1. Connect to a remote repository
 	ctx := context.Background()
-	layerDesc, err := pushBlob(ctx, ocispec.MediaTypeImageLayer, layerBlob, exampleMemoryStore) // push layer blob
+	reg := "myregistry.example.com"
+	repo, err := remote.NewRepository(reg + "/myrepo")
 	if err != nil {
 		panic(err)
 	}
-	configBlob := []byte("Hello config")
-	configDesc, err := pushBlob(ctx, ocispec.MediaTypeImageConfig, configBlob, exampleMemoryStore) // push config blob
+	// Note: The below code can be omitted if authentication is not required
+	repo.Client = &auth.Client{
+		Client: retry.DefaultClient,
+		Cache:  auth.DefaultCache,
+		Credential: auth.StaticCredential(reg, auth.Credential{
+			Username: "username",
+			Password: "password",
+		}),
+	}
+
+	// 2. Copy from the remote repository to the file store
+	tag := "latest"
+	manifestDescriptor, err := oras.Copy(ctx, repo, tag, fs, tag, oras.DefaultCopyOptions)
 	if err != nil {
 		panic(err)
 	}
-	manifestBlob, err := generateManifestContent(configDesc, layerDesc) // generate a image manifest
-	if err != nil {
-		panic(err)
-	}
-	manifestDesc, err := pushBlob(ctx, ocispec.MediaTypeImageManifest, manifestBlob, exampleMemoryStore) // push manifest blob
-	if err != nil {
-		panic(err)
-	}
-	err = exampleMemoryStore.Tag(ctx, manifestDesc, exampleTag)
+	fmt.Println("manifest descriptor:", manifestDescriptor)
+}
+
+func Example_pullImageFromRemoteRepository() {
+	// 0. Create an OCI layout store
+	store, err := oci.New("/tmp/oci-layout-root")
 	if err != nil {
 		panic(err)
 	}
 
-	// Setup example remote target
-	httpsServer := httptest.NewTLSServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		p := r.URL.Path
-		m := r.Method
-		switch {
-		case strings.Contains(p, "/blobs/uploads/") && m == "POST":
-			w.Header().Set("Content-Type", ocispec.MediaTypeImageManifest)
-			w.Header().Set("Location", p+exampleUploadUUid)
-			w.WriteHeader(http.StatusAccepted)
-		case strings.Contains(p, "/blobs/uploads/"+exampleUploadUUid) && m == "GET":
-			w.WriteHeader(http.StatusCreated)
-		case strings.Contains(p, "/manifests/"+string(exampleSignatureManifestDescriptor.Digest)):
-			w.Header().Set("Content-Type", spec.MediaTypeArtifactManifest)
-			w.Header().Set("Docker-Content-Digest", string(exampleSignatureManifestDescriptor.Digest))
-			w.Header().Set("Content-Length", strconv.Itoa(len(exampleSignatureManifest)))
-			w.Write(exampleSignatureManifest)
-		case strings.Contains(p, "/manifests/latest") && m == "PUT":
-			w.WriteHeader(http.StatusCreated)
-		case strings.Contains(p, "/manifests/"+string(exampleManifestDescriptor.Digest)),
-			strings.Contains(p, "/manifests/latest") && m == "HEAD":
-			w.Header().Set("Content-Type", spec.MediaTypeArtifactManifest)
-			w.Header().Set("Docker-Content-Digest", string(exampleManifestDescriptor.Digest))
-			w.Header().Set("Content-Length", strconv.Itoa(len(exampleManifest)))
-			if m == "GET" {
-				w.Write(exampleManifest)
-			}
-		case strings.Contains(p, "/v2/source/referrers/"):
-			var referrers []ocispec.Descriptor
-			if p == "/v2/source/referrers/"+exampleManifestDescriptor.Digest.String() {
-				referrers = []ocispec.Descriptor{exampleSignatureManifestDescriptor}
-			}
-			result := ocispec.Index{
-				Versioned: specs.Versioned{
-					SchemaVersion: 2, // historical value. does not pertain to OCI or docker version
-				},
-				MediaType: ocispec.MediaTypeImageIndex,
-				Manifests: referrers,
-			}
-			if err := json.NewEncoder(w).Encode(result); err != nil {
-				panic(err)
-			}
-		case strings.Contains(p, "/manifests/") && (m == "HEAD" || m == "GET"):
-			w.Header().Set("Content-Type", ocispec.MediaTypeImageManifest)
-			w.Header().Set("Docker-Content-Digest", string(manifestDesc.Digest))
-			w.Header().Set("Content-Length", strconv.Itoa(len([]byte(manifestBlob))))
-			w.Write([]byte(manifestBlob))
-		case strings.Contains(p, "/blobs/") && (m == "GET" || m == "HEAD"):
-			arr := strings.Split(p, "/")
-			digest := arr[len(arr)-1]
-			var desc ocispec.Descriptor
-			var content []byte
-			switch digest {
-			case layerDesc.Digest.String():
-				desc = layerDesc
-				content = layerBlob
-			case configDesc.Digest.String():
-				desc = configDesc
-				content = configBlob
-			case manifestDesc.Digest.String():
-				desc = manifestDesc
-				content = manifestBlob
-			}
-			w.Header().Set("Content-Type", desc.MediaType)
-			w.Header().Set("Docker-Content-Digest", digest)
-			w.Header().Set("Content-Length", strconv.Itoa(len([]byte(content))))
-			w.Write([]byte(content))
-		case strings.Contains(p, "/manifests/") && m == "PUT":
-			w.WriteHeader(http.StatusCreated)
+	// 1. Connect to a remote repository
+	ctx := context.Background()
+	reg := "myregistry.example.com"
+	repo, err := remote.NewRepository(reg + "/myrepo")
+	if err != nil {
+		panic(err)
+	}
+	// Note: The below code can be omitted if authentication is not required
+	repo.Client = &auth.Client{
+		Client: retry.DefaultClient,
+		Cache:  auth.DefaultCache,
+		Credential: auth.StaticCredential(reg, auth.Credential{
+			Username: "username",
+			Password: "password",
+		}),
+	}
+
+	// 2. Copy from the remote repository to the OCI layout store
+	tag := "latest"
+	manifestDescriptor, err := oras.Copy(ctx, repo, tag, store, tag, oras.DefaultCopyOptions)
+	if err != nil {
+		panic(err)
+	}
+	fmt.Println("manifest descriptor:", manifestDescriptor)
+}
+
+func Example_pushFilesToRemoteRepository() {
+	// 0. Create a file store
+	fs, err := file.New("/tmp/")
+	if err != nil {
+		panic(err)
+	}
+	defer fs.Close()
+	ctx := context.Background()
+
+	// 1. Add files to a file store
+	mediaType := "example/file"
+	fileNames := []string{"/tmp/myfile"}
+	fileDescriptors := make([]v1.Descriptor, 0, len(fileNames))
+	for _, name := range fileNames {
+		fileDescriptor, err := fs.Add(ctx, name, mediaType, "")
+		if err != nil {
+			panic(err)
 		}
+		fileDescriptors = append(fileDescriptors, fileDescriptor)
+		fmt.Printf("file descriptor for %s: %v\n", name, fileDescriptor)
+	}
 
-	}))
-	defer httpsServer.Close()
-	u, err := url.Parse(httpsServer.URL)
+	// 2. Pack the files and tag the packed manifest
+	artifactType := "example/files"
+	manifestDescriptor, err := oras.Pack(ctx, fs, artifactType, fileDescriptors, oras.DefaultPackOptions)
 	if err != nil {
 		panic(err)
 	}
-	remoteHost = u.Host
-	http.DefaultTransport = httpsServer.Client().Transport
+	fmt.Println("manifest descriptor:", manifestDescriptor)
 
-	os.Exit(m.Run())
-}
-
-func ExampleCopy_remoteToRemote() {
-	reg, err := remote.NewRegistry(remoteHost)
-	if err != nil {
-		panic(err) // Handle error
-	}
-	ctx := context.Background()
-	src, err := reg.Repository(ctx, "source")
-	if err != nil {
-		panic(err) // Handle error
-	}
-	dst, err := reg.Repository(ctx, "target")
-	if err != nil {
-		panic(err) // Handle error
-	}
-
-	tagName := "latest"
-	desc, err := oras.Copy(ctx, src, tagName, dst, tagName, oras.DefaultCopyOptions)
-	if err != nil {
-		panic(err) // Handle error
-	}
-	fmt.Println(desc.Digest)
-
-	// Output:
-	// sha256:7cbb44b44e8ede5a89cf193db3f5f2fd019d89697e6b87e8ed2589e60649b0d1
-}
-
-func ExampleCopy_remoteToLocal() {
-	reg, err := remote.NewRegistry(remoteHost)
-	if err != nil {
-		panic(err) // Handle error
-	}
-
-	ctx := context.Background()
-	src, err := reg.Repository(ctx, "source")
-	if err != nil {
-		panic(err) // Handle error
-	}
-	dst := memory.New()
-
-	tagName := "latest"
-	desc, err := oras.Copy(ctx, src, tagName, dst, tagName, oras.DefaultCopyOptions)
-	if err != nil {
-		panic(err) // Handle error
-	}
-	fmt.Println(desc.Digest)
-
-	// Output:
-	// sha256:7cbb44b44e8ede5a89cf193db3f5f2fd019d89697e6b87e8ed2589e60649b0d1
-}
-
-func ExampleCopy_localToLocal() {
-	src := exampleMemoryStore
-	dst := memory.New()
-
-	tagName := "latest"
-	ctx := context.Background()
-	desc, err := oras.Copy(ctx, src, tagName, dst, tagName, oras.DefaultCopyOptions)
-	if err != nil {
-		panic(err) // Handle error
-	}
-	fmt.Println(desc.Digest)
-
-	// Output:
-	// sha256:7cbb44b44e8ede5a89cf193db3f5f2fd019d89697e6b87e8ed2589e60649b0d1
-}
-
-func ExampleCopy_localToOciFile() {
-	src := exampleMemoryStore
-	tempDir, err := os.MkdirTemp("", "oras_oci_example_*")
-	if err != nil {
-		panic(err) // Handle error
-	}
-	defer os.RemoveAll(tempDir)
-	dst, err := oci.New(tempDir)
-	if err != nil {
-		panic(err) // Handle error
-	}
-
-	tagName := "latest"
-	ctx := context.Background()
-	desc, err := oras.Copy(ctx, src, tagName, dst, tagName, oras.DefaultCopyOptions)
-	if err != nil {
-		panic(err) // Handle error
-	}
-	fmt.Println(desc.Digest)
-
-	// Output:
-	// sha256:7cbb44b44e8ede5a89cf193db3f5f2fd019d89697e6b87e8ed2589e60649b0d1
-}
-
-func ExampleCopy_localToRemote() {
-	src := exampleMemoryStore
-	reg, err := remote.NewRegistry(remoteHost)
-	if err != nil {
-		panic(err) // Handle error
-	}
-	ctx := context.Background()
-	dst, err := reg.Repository(ctx, "target")
-	if err != nil {
-		panic(err) // Handle error
-	}
-
-	tagName := "latest"
-	desc, err := oras.Copy(ctx, src, tagName, dst, tagName, oras.DefaultCopyOptions)
-	if err != nil {
-		panic(err) // Handle error
-	}
-	fmt.Println(desc.Digest)
-
-	// Output:
-	// sha256:7cbb44b44e8ede5a89cf193db3f5f2fd019d89697e6b87e8ed2589e60649b0d1
-}
-
-// Example_copyArtifactManifestRemoteToLocal gives an example of copying
-// an artifact manifest from a remote repository to local.
-func Example_copyArtifactManifestRemoteToLocal() {
-	src, err := remote.NewRepository(fmt.Sprintf("%s/source", remoteHost))
-	if err != nil {
-		panic(err)
-	}
-	dst := memory.New()
-	ctx := context.Background()
-
-	exampleDigest := "sha256:70c29a81e235dda5c2cebb8ec06eafd3cca346cbd91f15ac74cefd98681c5b3d"
-	descriptor, err := src.Resolve(ctx, exampleDigest)
-	if err != nil {
-		panic(err)
-	}
-	err = oras.CopyGraph(ctx, src, dst, descriptor, oras.DefaultCopyGraphOptions)
-	if err != nil {
+	tag := "latest"
+	if err = fs.Tag(ctx, manifestDescriptor, tag); err != nil {
 		panic(err)
 	}
 
-	// verify that the artifact manifest described by the descriptor exists in dst
-	contentExists, err := dst.Exists(ctx, descriptor)
+	// 3. Connect to a remote repository
+	reg := "myregistry.example.com"
+	repo, err := remote.NewRepository(reg + "/myrepo")
 	if err != nil {
 		panic(err)
 	}
-	fmt.Println(contentExists)
+	// Note: The below code can be omitted if authentication is not required
+	repo.Client = &auth.Client{
+		Client: retry.DefaultClient,
+		Cache:  auth.DefaultCache,
+		Credential: auth.StaticCredential(reg, auth.Credential{
+			Username: "username",
+			Password: "password",
+		}),
+	}
 
-	// Output:
-	// true
-}
-
-// Example_extendedCopyArtifactAndReferrersRemoteToLocal gives an example of
-// copying an artifact along with its referrers from a remote repository to local.
-func Example_extendedCopyArtifactAndReferrersRemoteToLocal() {
-	src, err := remote.NewRepository(fmt.Sprintf("%s/source", remoteHost))
+	// 3. Copy from the file store to the remote repository
+	_, err = oras.Copy(ctx, fs, tag, repo, tag, oras.DefaultCopyOptions)
 	if err != nil {
 		panic(err)
 	}
-	dst := memory.New()
-	ctx := context.Background()
-
-	tagName := "latest"
-	// ExtendedCopy will copy the artifact tagged by "latest" along with all of its
-	// referrers from src to dst.
-	desc, err := oras.ExtendedCopy(ctx, src, tagName, dst, tagName, oras.DefaultExtendedCopyOptions)
-	if err != nil {
-		panic(err)
-	}
-
-	fmt.Println(desc.Digest)
-	// Output:
-	// sha256:f396bc4d300934a39ca28ab0d5ac8a3573336d7d63c654d783a68cd1e2057662
 }
