--- larger_example_test.go
@@ -1,162 +1,0 @@
-/*
-Copyright The ORAS Authors.
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package oras_test
-
-import (
-	"context"
-	"fmt"
-
-	v1 "github.com/opencontainers/image-spec/specs-go/v1"
-	"oras.land/oras-go/v2"
-	"oras.land/oras-go/v2/content"
-	"oras.land/oras-go/v2/content/file"
-	"oras.land/oras-go/v2/content/oci"
-	"oras.land/oras-go/v2/registry/remote"
-	"oras.land/oras-go/v2/registry/remote/auth"
-	"oras.land/oras-go/v2/registry/remote/retry"
-)
-
-func Example_pullFilesFromRemoteRepository() {
-	// 0. Create a file store
-	fs, err := file.New("/tmp/")
-	if err != nil {
-		panic(err)
-	}
-	defer fs.Close()
-
-	// 1. Connect to a remote repository
-	ctx := context.Background()
-	reg := "myregistry.example.com"
-	repo, err := remote.NewRepository(reg + "/myrepo")
-	if err != nil {
-		panic(err)
-	}
-	// Note: The below code can be omitted if authentication is not required
-	repo.Client = &auth.Client{
-		Client: retry.DefaultClient,
-		Cache:  auth.DefaultCache,
-		Credential: auth.StaticCredential(reg, auth.Credential{
-			Username: "username",
-			Password: "password",
-		}),
-	}
-
-	// 2. Copy from the remote repository to the file store
-	tag := "latest"
-	manifestDescriptor, err := oras.Copy(ctx, repo, tag, fs, tag, oras.DefaultCopyOptions)
-	if err != nil {
-		panic(err)
-	}
-	fmt.Println("manifest descriptor:", manifestDescriptor)
-}
-
-func Example_pullImageFromRemoteRepository() {
-	// 0. Create an OCI layout store
-	store, err := oci.New("/tmp/oci-layout-root")
-	if err != nil {
-		panic(err)
-	}
-
-	// 1. Connect to a remote repository
-	ctx := context.Background()
-	reg := "myregistry.example.com"
-	repo, err := remote.NewRepository(reg + "/myrepo")
-	if err != nil {
-		panic(err)
-	}
-	// Note: The below code can be omitted if authentication is not required
-	repo.Client = &auth.Client{
-		Client: retry.DefaultClient,
-		Cache:  auth.DefaultCache,
-		Credential: auth.StaticCredential(reg, auth.Credential{
-			Username: "username",
-			Password: "password",
-		}),
-	}
-
-	// 2. Copy from the remote repository to the OCI layout store
-	tag := "latest"
-	manifestDescriptor, err := oras.Copy(ctx, repo, tag, store, tag, oras.DefaultCopyOptions)
-	if err != nil {
-		panic(err)
-	}
-	fmt.Println("manifest pulled:", manifestDescriptor.Digest, manifestDescriptor.MediaType)
-
-	// 3. Fetch from OCI layout store to verify
-	fetched, err := content.FetchAll(ctx, store, manifestDescriptor)
-	if err != nil {
-		panic(err)
-	}
-	fmt.Printf("manifest content:\n%s", fetched)
-}
-
-func Example_pushFilesToRemoteRepository() {
-	// 0. Create a file store
-	fs, err := file.New("/tmp/")
-	if err != nil {
-		panic(err)
-	}
-	defer fs.Close()
-	ctx := context.Background()
-
-	// 1. Add files to a file store
-	mediaType := "example/file"
-	fileNames := []string{"/tmp/myfile"}
-	fileDescriptors := make([]v1.Descriptor, 0, len(fileNames))
-	for _, name := range fileNames {
-		fileDescriptor, err := fs.Add(ctx, name, mediaType, "")
-		if err != nil {
-			panic(err)
-		}
-		fileDescriptors = append(fileDescriptors, fileDescriptor)
-		fmt.Printf("file descriptor for %s: %v\n", name, fileDescriptor)
-	}
-
-	// 2. Pack the files and tag the packed manifest
-	artifactType := "example/files"
-	manifestDescriptor, err := oras.Pack(ctx, fs, artifactType, fileDescriptors, oras.DefaultPackOptions)
-	if err != nil {
-		panic(err)
-	}
-	fmt.Println("manifest descriptor:", manifestDescriptor)
-
-	tag := "latest"
-	if err = fs.Tag(ctx, manifestDescriptor, tag); err != nil {
-		panic(err)
-	}
-
-	// 3. Connect to a remote repository
-	reg := "myregistry.example.com"
-	repo, err := remote.NewRepository(reg + "/myrepo")
-	if err != nil {
-		panic(err)
-	}
-	// Note: The below code can be omitted if authentication is not required
-	repo.Client = &auth.Client{
-		Client: retry.DefaultClient,
-		Cache:  auth.DefaultCache,
-		Credential: auth.StaticCredential(reg, auth.Credential{
-			Username: "username",
-			Password: "password",
-		}),
-	}
-
-	// 3. Copy from the file store to the remote repository
-	_, err = oras.Copy(ctx, fs, tag, repo, tag, oras.DefaultCopyOptions)
-	if err != nil {
-		panic(err)
-	}
-}
